PR81635: Use chrecs to help find related data refs
The first loop in the testcase regressed after my recent changes to
dr_analyze_innermost.  Previously we would treat "i" as an iv even
for bb analysis and end up with:

   DR_BASE_ADDRESS: p or q
   DR_OFFSET: 0
   DR_INIT: 0 or 4
   DR_STEP: 16

We now always keep the step as 0 instead, so for an int "i" we'd have:

   DR_BASE_ADDRESS: p or q
   DR_OFFSET: i
   DR_INIT: 0 or 4
   DR_STEP: 0

This is also what we'd like to have for the unsigned "i", but the
problem is that strip_constant_offset thinks that the "i + 1" in
"(intptr_t) (i + 1)" could wrap and so doesn't peel off the "+ 1".
The [i + 1] accesses therefore have a DR_OFFSET equal to the SSA
name that holds i + 1, meaning that the accesses no longer seem
to be related to the [i] ones.

There seem to be two main uses of DR_OFFSET and DR_INIT.  One type
expects DR_OFFSET and DR_INIT to be generic expressions whose sum
gives the initial offset from DR_BASE_ADDRESS.  The other type uses
the pair (DR_BASE_ADDRESS, DR_OFFSET) to identify related data
references, with the distance between their start addresses being
the difference between their DR_INITs.  For this second type, the
exact form of DR_OFFSET doesn't really matter, and the more it is
able to canonicalise the non-constant offset, the better.

The patch fixes the PR by adding another offset/init pair to the
innermost loop behaviour for this second group.  The new pair use chrecs
rather than generic exprs for the offset part, with the chrec being
analysed in the innermost lopp for which the offset isn't invariant.
This allows us to vectorise the second function in the testcase
as well, which wasn't possible before the earlier patch.

2017-08-22  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	PR tree-optimization/81635
	* tree-data-ref.h (innermost_loop_behavior): Remove init field.
	Add var_offset and const_offset fields.  Rename offset_alignment
	to var_offset_alignment.
	(DR_INIT): Delete.
	(DR_CONST_OFFSET, DR_VAR_OFFSET): New macros.
	(DR_OFFSET_ALIGNMENT): Replace with...
	(DR_VAR_OFFSET_ALIGNMENT): ...this new macro.
	(dr_analyze_innermost): Add a gimple * argument.
	(dr_equal_offsets_p): Delete.
	(dr_var_offsets_equal_p, dr_var_offsets_compare): Declare.
	* tree-vectorizer.h (STMT_VINFO_DR_INIT): Delete.
	(STMT_VINFO_DR_VAR_OFFSET, STMT_VINFO_DR_CONST_OFFSET): New macros.
	(STMT_VINFO_DR_OFFSET_ALIGNMENT): Replace with...
	(STMT_VINFO_DR_VAR_OFFSET_ALIGNMENT): ...this new macro.
	* tree.c (tree_ctz): Handle POLYNOMIAL_CHREC.
	* tree-data-ref.c: Include tree-ssa-loop-ivopts.h.
	(split_constant_offset): Handle POLYNOMIAL_CHREC.
	(analyze_offset_scev): New function.
	(dr_analyze_innermost): Add a gimple * statement.  Update after
	changes to innermost_behavior.  Initialize var_offset and
	const_offset.
	(create_data_ref): Update call to dr_analyze_innermost.
	Update dump after changes to innermost_behavior.
	(operator ==): Use dr_var_offsets_equal_p and compare the
	DR_CONST_OFFSETs.
	(prune_runtime_alias_test_list): Likewise.
	(comp_dr_with_seg_len_pair): Use dr_var_offsets_compare and compare
	the DR_CONST_OFFSETs.
	(create_intersect_range_checks): Use DR_OFFSET without adding
	DR_INIT.
	(dr_equal_offsets_p1, dr_equal_offsets_p): Delete.
	(dr_alignment): Use const_offset instead of init, var_offset
	instead of offset, and var_offset_alignment instead of
	offset_alignment.
	* tree-if-conv.c (innermost_loop_behavior_hash::hash): Don't
	test the init field.
	(innermost_loop_behavior_hash::equal): Likewise.
	(ifcvt_memrefs_wont_trap): Likewise.
	(if_convertible_loop_p_1): Likewise.
	* tree-loop-distribution.c (build_addr_arg_loc): Use DR_OFFSET
	without adding DR_INIT.
	(build_rdg_partition_for_vertex): Don't check DR_INIT.
	(share_memory_accesses): Likewise.
	(pg_add_dependence_edges): Likewise.
	(compute_alias_check_pairs): Use dr_var_offsets_compare.
	* tree-predcom.c (aff_combination_dr_offset): Use DR_OFFSET without
	adding DR_INIT.
	(determine_offset): Likewise.
	(valid_initializer_p): Likewise.
	(find_looparound_phi): Update call to dr_analyze_innermost.
	(ref_at_iteration): Use split_constant_offset to split the offset.
	* tree-vect-data-refs.c (vect_compute_data_ref_alignment): Use
	const_offset instead of init and var_offset_alignment instead of
	offset_alignment.
	(vect_find_same_alignment_drs): Use dr_var_offsets_compare and
	compare the DR_CONST_OFFSETs.
	(dr_group_sort_cmp): Likewise.
	(vect_analyze_group_access_1): Use DR_CONST_OFFSET instead of DR_INIT.
	(vect_no_alias_p): Likewise.
	(vect_analyze_data_ref_accesses): Use dr_var_offsets_equal_p and
	compare the DR_CONST_OFFSETs.
	(vect_prune_runtime_alias_test_list): Use dr_var_offsets_compare.
	(vect_analyze_data_refs): Don't check DR_INIT and use DR_OFFSET
	without adding DR_INIT.  Use DR_VAR_OFFSET_ALIGNMENT instead of
	DR_OFFSET_ALIGNMENT.  Update call to dr_analyze_innermost, and
	update subsequent dump.  Initialize DR_VAR_OFFSET and
	DR_CONST_OFFSET when filling in the dr manually.
	(vect_create_addr_base_for_vector_ref): Use DR_OFFSET without
	adding DR_INIT.
	* tree-vect-stmts.c (vectorizable_store): Likewise.
	(vectorizable_load): Likewise.  Use DR_CONST_OFFSET instead
	of DR_INIT.

gcc/testsuite/
	PR tree-optimization/81635
	* gcc.dg/vect/bb-slp-pr81635.c: New test.